benchmark/requirements.txt
numpy==1.24.2
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
tensorflow-cpu==2.11.0
matplotlib==3.7.0
tqdm==4.64.1
onnx==1.12.0
onnxruntime==1.14.1